removed decoding, reordered utils functions
